requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==3.2
django-extensions==3.2.1
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
executing==1.2.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==1.7.1
pytz==2022.7
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.8.0
wcwidth==0.2.5
